segregate certificate display in print_certificates
